requirements.txt
pandas==2.2.2
python-docx==1.1.2
pillow==10.4.0
pdf2docx==0.5.8
PyPDF2==3.0.1
tabula-py==2.9.3
rarfile==4.2
JPype1==1.5.0
openpyxl==3.1.2
google-auth==2.35.0
gspread==6.1.3
inflect==7.4.0
appscript==1.3.0
docx2pdf==0.1.8
tqdm==4.66.6
weasyprint==63.0
zopfli==0.2.3.post1
Pyphen==0.15.0
brotli==1.1.0
cssselect2==0.7.0
pydyf==0.11.0
tinycss2==1.4.0
tinyhtml5==2.0.0
bs4==0.0.2





